gtest fin

+++ calculator/calculator.h
@@ -4,5 +4,5 @@
 int addnum(int,int);
 int subnum(int,int);
 int mulnum(int,int);
-int divnum(int,int);
+int divnum(double,double);
 #endif
